+++ src/dns.rs
@@ -1,49 +1,39 @@
 use indexmap::IndexMap;
-use std::cell::RefCell;
 use std::net::SocketAddr;
 
 pub struct Dns {
-    inner: RefCell<Inner>,
-}
-
-struct Inner {
     next: u16,
     names: IndexMap<String, u16>,
 }
 
 pub trait ToSocketAddr {
-    fn to_socket_addr(&self, dns: &Dns) -> SocketAddr;
+    fn to_socket_addr(&self, dns: &mut Dns) -> SocketAddr;
 }
 
 impl Dns {
     pub(crate) fn new() -> Dns {
         Dns {
-            inner: RefCell::new(Inner {
-                next: 1,
-                names: IndexMap::new(),
-            }),
+            next: 1,
+            names: IndexMap::new(),
         }
     }
 
-    pub(crate) fn lookup(&self, addr: impl ToSocketAddr) -> SocketAddr {
+    pub(crate) fn lookup(&mut self, addr: impl ToSocketAddr) -> SocketAddr {
         addr.to_socket_addr(self)
     }
 }
 
 impl ToSocketAddr for String {
-    fn to_socket_addr(&self, dns: &Dns) -> SocketAddr {
+    fn to_socket_addr(&self, dns: &mut Dns) -> SocketAddr {
         (&self[..]).to_socket_addr(dns)
     }
 }
 
 impl<'a> ToSocketAddr for &'a str {
-    fn to_socket_addr(&self, dns: &Dns) -> SocketAddr {
-        let mut inner = dns.inner.borrow_mut();
-        let inner = &mut *inner;
-
-        let host = inner.names.entry(self.to_string()).or_insert_with(|| {
-            let next = inner.next;
-            inner.next += 1;
+    fn to_socket_addr(&self, dns: &mut Dns) -> SocketAddr {
+        let host = dns.names.entry(self.to_string()).or_insert_with(|| {
+            let next = dns.next;
+            dns.next += 1;
             next
         });
 
@@ -55,7 +45,7 @@ impl<'a> ToSocketAddr for &'a str {
 }
 
 impl ToSocketAddr for SocketAddr {
-    fn to_socket_addr(&self, _: &Dns) -> SocketAddr {
+    fn to_socket_addr(&self, _: &mut Dns) -> SocketAddr {
         *self
     }
 }
