+++ src/world.rs
@@ -1,16 +1,14 @@
-use crate::top::Latency;
-use crate::{inbox, version, Dns, ToSocketAddr, Topology};
+use crate::{config, version, Dns, Host, ToSocketAddr, Topology};
 
 use indexmap::IndexMap;
 use rand::RngCore;
 use scoped_tls::scoped_thread_local;
 use std::any::Any;
 use std::cell::RefCell;
-use std::collections::VecDeque;
 use std::net::SocketAddr;
 use std::sync::Arc;
 use tokio::sync::Notify;
-use tokio::time::{Duration, Instant};
+use tokio::time::Instant;
 
 /// Tracks all the state for the simulated world.
 pub(crate) struct World {
@@ -18,7 +16,7 @@ pub(crate) struct World {
     hosts: IndexMap<SocketAddr, Host>,
 
     /// Tracks how each host is connected to each other.
-    topology: Topology,
+    pub(crate) topology: Topology,
 
     /// Maps hostnames to socket addresses.
     dns: Dns,
@@ -35,10 +33,10 @@ scoped_thread_local!(static CURRENT: RefCell<World>);
 
 impl World {
     /// Initialize a new world
-    pub(crate) fn new(latency: Latency, rng: Box<dyn RngCore>) -> World {
+    pub(crate) fn new(link: config::Link, rng: Box<dyn RngCore>) -> World {
         World {
             hosts: IndexMap::new(),
-            topology: Topology::new(latency),
+            topology: Topology::new(link),
             dns: Dns::new(),
             current: None,
             rng,
@@ -66,14 +64,16 @@ impl World {
         self.hosts.get_mut(&addr).expect("host missing")
     }
 
-    /// Return a mutable reference to the currently running host state
-    pub(crate) fn current_mut(&mut self) -> Option<&mut Host> {
-        let addr = self.current.expect("no current host");
-        self.hosts.get_mut(&addr)
-    }
-
     pub(crate) fn lookup(&mut self, host: impl ToSocketAddr) -> SocketAddr {
         self.dns.lookup(host)
+    }
+
+    pub(crate) fn partition(&mut self, a: SocketAddr, b: SocketAddr) {
+        self.topology.partition(a, b);
+    }
+
+    pub(crate) fn repair(&mut self, a: SocketAddr, b: SocketAddr) {
+        self.topology.repair(a, b);
     }
 
     /// Register a new host with the simulation
@@ -89,7 +89,7 @@ impl World {
         }
 
         // Initialize host state
-        self.hosts.insert(addr, Host::new(addr, epoch, notify));
+        self.hosts.insert(addr, Host::new(epoch, notify));
     }
 
     /// Send a message between two hosts
@@ -107,101 +107,5 @@ impl World {
     /// Tick a host
     pub(crate) fn tick(&mut self, addr: SocketAddr, now: Instant) {
         self.hosts.get_mut(&addr).expect("missing host").tick(now);
-    }
-}
-
-// ===== TODO: Move this to host.rs =====
-
-/// A host in the simulated network
-pub(crate) struct Host {
-    /// Host's socket address
-    pub(crate) addr: SocketAddr,
-
-    /// Messages in-flight to the host. Some of these may still be "on the
-    /// network".
-    inbox: IndexMap<SocketAddr, VecDeque<inbox::Envelope>>,
-
-    /// Signaled when a message becomes available to receive
-    notify: Arc<Notify>,
-
-    /// Current instant at the host
-    pub(crate) now: Instant,
-
-    /// Instant at which the host began
-    epoch: Instant,
-
-    /// Current host version. This is incremented each time a message is received.
-    version: u64,
-}
-
-impl Host {
-    pub(crate) fn new(addr: SocketAddr, epoch: Instant, notify: Arc<Notify>) -> Host {
-        Host {
-            addr,
-            inbox: IndexMap::new(),
-            notify,
-            now: epoch,
-            epoch,
-            version: 0,
-        }
-    }
-
-    pub(crate) fn send(&mut self, src: version::Dot, delay: Duration, message: Box<dyn Any>) {
-        let deliver_at = self.now + delay;
-
-        self.inbox
-            .entry(src.host)
-            .or_default()
-            .push_back(inbox::Envelope {
-                src,
-                deliver_at,
-                message,
-            });
-
-        self.notify.notify_one();
-    }
-
-    pub(crate) fn recv(&mut self) -> Option<inbox::Envelope> {
-        let now = Instant::now();
-
-        for deque in self.inbox.values_mut() {
-            match deque.front() {
-                Some(inbox::Envelope { deliver_at, .. }) if *deliver_at <= now => {
-                    self.version += 1;
-                    return deque.pop_front();
-                }
-                _ => {
-                    // Fall through to the notify
-                }
-            }
-        }
-
-        None
-    }
-
-    pub(crate) fn recv_from(&mut self, src: SocketAddr) -> Option<inbox::Envelope> {
-        let now = Instant::now();
-        let deque = self.inbox.entry(src).or_default();
-
-        match deque.front() {
-            Some(inbox::Envelope { deliver_at, .. }) if *deliver_at <= now => {
-                self.version += 1;
-                deque.pop_front()
-            }
-            _ => None,
-        }
-    }
-
-    pub(crate) fn tick(&mut self, now: Instant) {
-        self.now = now;
-
-        for deque in self.inbox.values() {
-            if let Some(inbox::Envelope { deliver_at, .. }) = deque.front() {
-                if *deliver_at <= now {
-                    self.notify.notify_one();
-                    return;
-                }
-            }
-        }
     }
 }
